+++ android/app/src/main/java/in/adaniports/portvision3d/MainActivity.java
@@ -1,0 +1,5 @@
+package in.adaniports.portvision3d;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}